Little tweaks to prevent accidents.
git-svn-id: http://svn.code.sf.net/p/postbooks/code/xtuple/trunk@3037 ea1a15ef-c238-0410-93cb-fa28e9f07733

+++ guiclient/scripts.cpp
@@ -139,12 +139,19 @@ void scripts::sEdit()
 
 void scripts::sDelete()
 {
-  q.prepare( "DELETE FROM script "
-             "WHERE (script_id=:script_id);" );
-  q.bindValue(":script_id", _script->id());
-  q.exec();
+  if ( QMessageBox::warning(this, tr("Delete Script?"),
+                            tr("<p>Are you sure that you want to completely "
+			       "delete the selected script?"),
+			    QMessageBox::Yes,
+			    QMessageBox::No | QMessageBox::Default) == QMessageBox::Yes)
+  {
+    q.prepare( "DELETE FROM script "
+               "WHERE (script_id=:script_id);" );
+    q.bindValue(":script_id", _script->id());
+    q.exec();
 
-  sFillList();
+    sFillList();
+  }
 }
 
 void scripts::sFillList()
